refactor: 서비스 이벤트 리스너 수정
