+++ src/main/java/kitchenpos/table/application/TableService.java
@@ -1,23 +1,16 @@
 package kitchenpos.table.application;
 
-import kitchenpos.common.GroupOrderTablesEvent;
-import kitchenpos.common.UngroupOrderTableEvent;
 import kitchenpos.common.ValidateOrderTableOrderStatusEvent;
-import kitchenpos.common.ValidateOrderTablesOrderStatusEvent;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.table.dto.OrderTableResponse;
-import kitchenpos.table.exception.InvalidOrderTableException;
 import kitchenpos.table.exception.OrderTableNotFoundException;
-import kitchenpos.table.exception.DuplicateOrderTableException;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -67,39 +60,5 @@ public class TableService {
         final OrderTable savedOrder = orderTableRepository.save(newOrderTable);
 
         return OrderTableResponse.from(savedOrder);
-    }
-
-    @EventListener
-    public void ungroup(final UngroupOrderTableEvent ungroupOrderTableEvent) {
-        final List<OrderTable> orderTables = orderTableRepository.findAllByTableGroupId(ungroupOrderTableEvent.getTableGroupId());
-        final List<Long> orderTableIds = orderTables.stream()
-                .map(it -> it.getId())
-                .collect(Collectors.toList());
-        eventPublisher.publishEvent(new ValidateOrderTablesOrderStatusEvent(orderTableIds));
-
-        orderTables.stream()
-                .forEach(orderTable -> orderTable.ungroup());
-    }
-
-    @EventListener
-    public void group(final GroupOrderTablesEvent groupOrderTablesEvent) {
-        final List<Long> orderTableIds = groupOrderTablesEvent.getOrderTableIds();
-        final List<OrderTable> orderTables = orderTableRepository.findByIdIn(orderTableIds);
-
-        validateOrderTable(orderTables, orderTableIds);
-
-        orderTables.stream().forEach(orderTable -> orderTable.changeTableGroup(groupOrderTablesEvent.getTableGroupId()));
-    }
-
-    private void validateOrderTable(final List<OrderTable> orderTables, final List<Long> orderTableIds) {
-        if (orderTables.size() != orderTableIds.size()) {
-            throw new InvalidOrderTableException();
-        }
-
-        for (final OrderTable savedOrderTable : orderTables) {
-            if (savedOrderTable.isEmpty() || Objects.nonNull(savedOrderTable.getTableGroupId())) {
-                throw new DuplicateOrderTableException();
-            }
-        }
     }
 }
